requirements.txt
absl-py==0.1.11
astor==0.6.2
backports.weakref==1.0.post1
bleach==1.5.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.0
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
Keras==2.1.5
Markdown==2.6.11
mock==2.0.0
numpy==1.14.2
pbr==3.1.1
Pillow==5.0.0
protobuf==3.5.2
PyYAML==3.12
scipy==1.0.1
six==1.11.0
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
Werkzeug==0.14.1
